Working on custom controls to replace xceed toolkit.

+++ WinHue3/Philips Hue/HueObjects/Common/IBaseProperties.cs
@@ -19,7 +19,7 @@
 
         string colormode { get; set; }
 
-        uint? transitiontime { get; set; }
+        ushort? transitiontime { get; set; }
 
         string alert { get; set; }
